Fixes for compliance with official kodi plugin repo travis test (Closes #88)
